TIP-236: Move user event to component
